+++ packages/elements/src/components/ui/action-dropdown/ActionDropdown.tsx
@@ -1,4 +1,5 @@
-import { faCaretDown } from "@fortawesome/free-solid-svg-icons/faCaretDown";
+import { faChevronDown } from "@fortawesome/free-solid-svg-icons/faChevronDown";
+import { faChevronUp } from "@fortawesome/free-solid-svg-icons/faChevronUp";
 import {
   Box,
   Clickable,
@@ -32,18 +33,29 @@ export const ActionDropdown: React.FC<ActionDropdownProps> = ({
   const [expanded, open, close] = useBoolean(false);
   const ref = useRef(null);
   useOnClickOutside(ref, close);
+
   const { colors } = useThemeFields(
     {
       colors: {
         textColor: theme.textColor,
         background: theme.background,
+        borderColor: theme.borderColor,
+        borderColorHover: theme.borderColorHover,
         dropdownBackground: theme.dropdownBackground
       }
     },
     [theme]
   );
+
+  const hoverBorder = `1px solid ${colors.borderColorHover}`;
+
   return (
-    <Box position={"relative"} display={"inline-block"} width={width}>
+    <Box
+      position={"relative"}
+      display={"inline-block"}
+      width={width}
+      innerRef={ref}
+    >
       {expanded && (
         <Column
           position={"absolute"}
@@ -51,43 +63,44 @@ export const ActionDropdown: React.FC<ActionDropdownProps> = ({
           left={0}
           right={0}
           background={colors.dropdownBackground}
-          borderColor={theme.borderColor}
+          borderColor={colors.borderColorHover}
           borderRadius={theme.borderRadius}
           borderWidth={1}
           borderStyle={"solid"}
         >
           <Row
             width={"100%"}
+            height={theme.height}
             justifyContent={"space-between"}
             alignItems={"center"}
-            height={theme.height}
             indent
           >
             <StandardText>{label}</StandardText>
-            <Icon icon={faCaretDown} />
+            <Icon icon={faChevronUp} size={12} />
           </Row>
-          {children}
+          <Column spacing={0.5}>{children}</Column>
         </Column>
       )}
       <Clickable onClick={open} disableFocusHighlight>
         <Box
-          borderColor={theme.borderColor}
+          borderColor={colors.borderColor}
+          hoverBorder={hoverBorder}
           borderRadius={theme.borderRadius}
           borderWidth={1}
           borderStyle={"solid"}
           background={colors.background}
           width={width}
+          height={theme.height}
         >
           <Row
             width={"100%"}
-            height={theme.height}
-            innerRef={ref}
+            height={"100%"}
             alignItems={"center"}
             justifyContent={"space-between"}
             indent
           >
             <StandardText>{label}</StandardText>
-            <Icon icon={faCaretDown} />
+            <Icon icon={expanded ? faChevronUp : faChevronDown} size={12} />
           </Row>
         </Box>
       </Clickable>