+++ packages/elements/src/components/ui/action-dropdown/ActionDropdownSeparator.tsx
@@ -1,0 +1,14 @@
+import { SeparatorLine, Spacing } from "@stenajs-webui/core";
+import * as React from "react";
+
+interface ActionDropdownSeparatorProps {}
+
+export const ActionDropdownSeparator: React.FC<
+  ActionDropdownSeparatorProps
+> = () => {
+  return (
+    <Spacing>
+      <SeparatorLine />
+    </Spacing>
+  );
+};