+++ packages/elements/src/components/ui/action-dropdown/ActionDropdownTheme.ts
@@ -4,13 +4,20 @@ export interface ActionDropdownTheme {
   background?: ThemeColorField | string;
   textColor?: ThemeColorField | string;
   borderColor?: ThemeColorField | string;
+  borderColorHover?: ThemeColorField | string;
   borderRadius?: string;
   dropdownBackground?: ThemeColorField | string;
+  iconColor: ThemeColorField | string;
+  iconColorDisabled: ThemeColorField | string;
+  iconColorHover: ThemeColorField | string;
   itemBackground?: ThemeColorField | string;
+  itemBackgroundDisabled?: ThemeColorField | string;
   itemBackgroundHover?: ThemeColorField | string;
   itemLabelColor?: ThemeColorField | string;
+  itemLabelColorDisabled?: ThemeColorField | string;
   itemLabelColorHover?: ThemeColorField | string;
   itemTextColor?: ThemeColorField | string;
+  itemTextColorDisabled?: ThemeColorField | string;
   itemTextColorHover?: ThemeColorField | string;
   height: string;
   itemHeight: string;
@@ -20,14 +27,21 @@ export const defaultActionDropdownTheme: ActionDropdownTheme = {
   background: "white",
   textColor: "primaryText",
   borderColor: "inputBorder",
+  borderColorHover: "inputBorderFocused",
   borderRadius: "4px",
   dropdownBackground: "white",
-  itemBackground: undefined,
-  itemBackgroundHover: "disabledBackground",
+  iconColor: "primaryText",
+  iconColorDisabled: "disabledText",
+  iconColorHover: "white",
+  itemBackground: "white",
+  itemBackgroundHover: "interactionBgPrimary",
+  itemBackgroundDisabled: "disabledBackground",
   itemLabelColor: "primaryText",
-  itemLabelColorHover: undefined,
+  itemLabelColorDisabled: "disabledText",
+  itemLabelColorHover: "white",
   itemTextColor: "primaryTextLight",
-  itemTextColorHover: undefined,
+  itemTextColorDisabled: "disabledTextLight",
+  itemTextColorHover: "white",
   height: "34px",
   itemHeight: "34px"
 };