+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 #### Next release
 
-* Coming soon... 
+* Add `ActionDropdown` component to `elements`.
 
 #### 0.0.35
 
 * Upgrade `date-fns` to `2.0.1`.
 * Add `show` option to `WeekNumber` config used in `CalenderMonth`
 * Add new highlight states to `DayStateHighlight`: `selectedStart` and `selectedEnd`. For compatibility reasons the selected days hove both the old `selected` AND `selectedStart` and/or `selectedEnd`
-* Add `borderColor` (and `borderColorDisabled`) to `ButtonTheme`
+* Add `borderColor` (and `borderColorDisabled`) to `ButtonTheme`.
 
 #### 0.0.34